Acadimic Career Done

+++ lib/models/Semester_logs_models/course_model.dart
@@ -3,7 +3,7 @@ enum Grades { A, A_Minus, B_Plus, B_Minus, C_Plus, C, D, F }
 class CourseModel {
   final String courseCode;
   final String courseName;
-  final int grade;
+  final double grade; // <-- changed from int to double
   final int creditHours;
   final double courseScore;
   String? gradeLetter;
@@ -11,7 +11,7 @@ class CourseModel {
   CourseModel({
     required this.courseCode,
     required this.courseName,
-    required this.grade,
+    required this.grade, // <-- changed from int to double
     required this.creditHours,
     required this.courseScore,
     this.gradeLetter,
